Alliance works without private data...
