+++ alliance/go.mod
@@ -1,0 +1,15 @@
+module github.com/stefanprisca/strategy-code/alliance
+
+go 1.12
+
+require (
+	github.com/gogo/protobuf v1.2.1
+	github.com/golang/protobuf v1.3.1
+	github.com/hyperledger/fabric v1.4.1
+	github.com/stefanprisca/strategy-code/tfc v0.0.0-20190523135922-1891d5cabd84
+	github.com/stefanprisca/strategy-protobufs v0.0.0-20190518070239-5e1d55254410
+	github.com/stretchr/testify v1.3.0
+	golang.org/x/crypto v0.0.0-20190513172903-22d7a77e9e5f // indirect
+	golang.org/x/net v0.0.0-20190522155817-f3200d17e092 // indirect
+	google.golang.org/grpc v1.21.0 // indirect
+)
