+++ cmd/alliance/go.mod
@@ -4,7 +4,7 @@ go 1.12
 
 require (
 	github.com/hyperledger/fabric v1.4.1
-	github.com/stefanprisca/strategy-code/tfc v0.0.0-20190523134131-0e565b72f206
+	github.com/stefanprisca/strategy-code/tfc v0.0.0-20190523135922-1891d5cabd84
 	golang.org/x/crypto v0.0.0-20190513172903-22d7a77e9e5f // indirect
 	golang.org/x/net v0.0.0-20190514140710-3ec191127204 // indirect
 )
