--- go.mod
@@ -1,11 +1,0 @@
-module github.com/stefanprisca/strategy-code
-
-go 1.12
-
-require (
-	github.com/golang/protobuf v1.3.1
-	github.com/hyperledger/fabric v1.4.1
-	github.com/stefanprisca/strategy-code/tfc v0.0.0-20190504143904-4cc3439a5c77
-	github.com/stefanprisca/strategy-protobufs v0.0.0-20190502133108-0006c391fc13
-	github.com/stretchr/testify v1.3.0
-)
