--- tfc/alliance/argsbuilder.go
@@ -1,1 +1,0 @@
-package alliance
